src: from /types moved boavizta types to the corresponding folder
