+++ src/types/index.ts
@@ -1,3 +1,3 @@
-export * from './boavizta';
+export * from '../lib/boavizta/types';
 export * from './ccf';
 export * from './common';
